Move StyleGroupsAlgorithm, F32TaskInput, StyleGroupsConfig into style_groups
These types are the algorithm-neutral entry points used by both the loose
and graph chunking algorithms (and by every chunking-context call site
through ChunkingConfig::style_groups_algorithm). Living in
style_groups_loose made the imports misleading and the module name
suggested the type was loose-specific.
