+++ crates/next-core/src/next_edge/context.rs
@@ -14,7 +14,7 @@ use turbopack_core::{
     free_var_references,
     issue::IssueSeverity,
     module_graph::{
-        binding_usage_info::OptionBindingUsageInfo, style_groups_loose::StyleGroupsAlgorithm,
+        binding_usage_info::OptionBindingUsageInfo, style_groups::StyleGroupsAlgorithm,
     },
 };
 use turbopack_css::chunk::CssChunkType;
