+++ crates/next-core/src/next_server/context.rs
@@ -23,7 +23,7 @@ use turbopack_core::{
     environment::{Environment, ExecutionEnvironment, NodeJsEnvironment, NodeJsVersion},
     issue::IssueSeverity,
     module_graph::{
-        binding_usage_info::OptionBindingUsageInfo, style_groups_loose::StyleGroupsAlgorithm,
+        binding_usage_info::OptionBindingUsageInfo, style_groups::StyleGroupsAlgorithm,
     },
     target::CompileTarget,
 };
