+++ turbopack/crates/turbopack-core/src/module_graph/mod.rs
@@ -35,9 +35,9 @@ use crate::{
         chunk_group_info::{ChunkGroupEntry, ChunkGroupInfo, compute_chunk_group_info},
         merged_modules::{MergedModuleInfo, compute_merged_modules},
         module_batches::{ModuleBatchesGraph, compute_module_batches},
-        style_groups::StyleGroups,
+        style_groups::{StyleGroups, StyleGroupsAlgorithm, StyleGroupsConfig},
         style_groups_graph::compute_style_groups_graph,
-        style_groups_loose::{StyleGroupsAlgorithm, StyleGroupsConfig, compute_style_groups},
+        style_groups_loose::compute_style_groups,
         traced_di_graph::TracedDiGraph,
     },
     reference::{ModuleReference, primary_chunkable_referenced_modules},
